# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for Air Social application.",
        "title": "Air Social API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/air-social/api/v1",
    "paths": {
...
        "/media/presigned-urls": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a list of presigned URLs and policy signatures for direct file upload to MinIO/S3.\n\n**Client Integration:**\n* **Target:** Send a `POST` request to the returned `upload_url`.\n* **Payload:** Use `multipart/form-data`. Append all fields from `form_data` (must be FIRST) followed by the file binary (must be LAST).\n* **Success:** Upon successful upload (HTTP 204), confirm the `object_key` with the Backend database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Get presigned upload URL",
                "parameters": [
                    {
                        "description": "File Metadata List",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_media.BulkPresignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upload Credentials List",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_transport_http_media.PresignResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.ValidationResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "air-social_internal_domain_media.UploadDomain": {
            "type": "string",
            "enum": [
                "users",
                "posts",
                "messages"
            ],
            "x-enum-varnames": [
                "DomainUser",
                "DomainPost",
                "DomainMessage"
            ]
        },
        "air-social_internal_domain_media.UploadFeature": {
            "type": "string",
            "enum": [
                "avatar",
                "cover",
                "feed_image",
                "feed_video",
                "voice_chat",
                "attachment"
            ],
            "x-enum-varnames": [
                "FeatureAvatar",
                "FeatureCover",
                "FeatureFeedImage",
                "FeatureFeedVideo",
                "FeatureVoiceChat",
                "FeatureAttachment"
            ]
        },
...
        "internal_transport_http_media.BulkPresignRequest": {
            "type": "object",
            "required": [
                "files"
            ],
            "properties": {
                "files": {
                    "type": "array",
                    "maxItems": 10,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/internal_transport_http_media.PresignRequest"
                    }
                }
            }
        },
        "internal_transport_http_media.PresignRequest": {
            "type": "object",
            "required": [
                "domain",
                "feature",
                "file_name",
                "file_size",
                "file_type"
            ],
            "properties": {
                "domain": {
                    "enum": [
                        "users",
                        "posts",
                        "groups",
                        "messages"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/air-social_internal_domain_media.UploadDomain"
                        }
                    ]
                },
                "feature": {
                    "enum": [
                        "avatar",
                        "cover",
                        "feed_image",
                        "feed_video",
                        "voice_chat",
                        "attachment"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/air-social_internal_domain_media.UploadFeature"
                        }
                    ]
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_media.PresignResponse": {
            "type": "object",
            "properties": {
                "expire_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "form_data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "object_key": {
                    "type": "string"
                },
                "public_url": {
                    "type": "string"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
        "pkg.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "pkg.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "pkg.ValidationResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.FieldError"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}